Factor out observations/starcats tests to a new test file
